[POC] Add 2nd config to compile bundeled Svelte components as Webcomponent library
SvelteKit can not directly compile to WebComponents as a Javascript
library. One has to feed assembled components through a 2nd build
step. This is what the 2nd config is for.

Bundle Webcomponent as Javascript library:

   $ npm run build-webcomponent

This will build a file to /dist-js/.

Credits to this GitHub issue: https://github.com/sveltejs/kit/issues/10320#issue-1789185724